+++ book-review.php
@@ -16,7 +16,7 @@
  * Plugin Name:       Book Review
  * Plugin URI:        http://wpreviewplugins.com/downloads/book-review/
  * Description:       Add book information such as title, author, publisher and cover photo to enhance your review posts.
- * Version:           2.3.0
+ * Version:           2.3.1
  * Author:            Donna Peplinskie
  * Author URI:        http://donnapeplinskie.com/
  * Text Domain:       book-review