+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Change Log #
 
+## 2.3.1 ##
+* New: Add new `book_review_links_meta` filter for returning links meta used in the *Book Info* meta box.
+* Fix: Change `book_review_links` filter to return links HTML.
+
 ## 2.3.0 ##
 * New: Add support for custom fields.
 * Tweak: Improved validation when saving settings and post meta.